refactor: make the vertical layout use the new way of using the writer

+++ crates/os/src/ui/writer.rs
@@ -134,7 +134,7 @@ pub fn clear() {
 }
 
 #[inline]
-pub fn clear_zone(from: (usize, usize), to: (usize, usize)) {
+pub fn clear_zone(from: (u64, u64), to: (u64, u64)) {
     let mut ui = get_ui();
 
     ui.clear_zone(from, to);
